added api_definitions
